mafgood_updated/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
Authlib==1.3.0
requests==2.31.0
python-dotenv==1.0.0
Flask
Authlib
requests
gunicorn
python-dotenv
gunicorn app:app --bind 0.0.0.0:$PORT
--testing 